rename locales to match woth iso
